backend/requirements.txt
django==3.2.9
psycopg2-binary==2.9.1
django-cors-headers==4.2.0
certifi==2023.7.22 
charset-normalizer==3.2.0 
idna==3.4 
requests==2.31.0 
urllib3==2.0.5
djangorestframework==3.14.0